refactor(processes): split lib.rs into per-concern modules
Carves the 1953-line lib.rs into 7 focused modules so each file fits
in one tool-read and an agent can grep-jump straight to the right
concern:

- types.rs            data types, errors, traits, shared helpers
- cancellation.rs     token + registry
- host.rs             ProcessHost + Subscription
- memory_store.rs     InMemory{Store,ResultStore}
- filesystem_store.rs Filesystem* + path/serde helpers
- wrappers.rs         Eventing + ResourceManaged decorators
- services.rs         ProcessServices + BackgroundProcessManager

lib.rs is now a thin module-decl + re-export hub so the public API
surface is visible at a glance.

No behavior change. Verified:
- cargo fmt
- cargo check -p ironclaw_processes
- cargo clippy -p ironclaw_processes --all-targets -- -D warnings
- cargo test -p ironclaw_processes (43/43 pass)
